Refactor core modules into echo_core
