--- belief/test
@@ -1,1 +1,0 @@
-
