+++ core/__init__.py
@@ -1,4 +1,6 @@
 from importlib import import_module
 import sys
-module = import_module('universalCode.core')
+
+# TODO: REMOVE after transition to echo_core.tools
+module = import_module('universalCode.echo_core.tools')
 sys.modules[__name__] = module
