--- docs/README.md
@@ -1,3 +1,0 @@
-# Documentation
-
-This folder contains memory logs, symbolic definitions, and RAIP-R specifications.
