+++ tests/belief_placeholder_test
@@ -1,0 +1,1 @@
+# TODO: REMOVE placeholder test file
