+++ universalCode/echo_core/tools/__init__.py
@@ -1,6 +1,8 @@
 from .belief_export_formatter import format_beliefs, HANDSHAKE_SIGNATURE
 from .triad_translator import extract_propositional_cores, display_cores_yaml
 from .memory_ingestor import ingest_entries, ingest_yaml
+from .raip_r_engine import RAIPREngine
+from .belief_validation_engine import BeliefValidationEngine
 
 __all__ = [
     "format_beliefs",
@@ -9,4 +11,6 @@ __all__ = [
     "display_cores_yaml",
     "ingest_entries",
     "ingest_yaml",
+    "RAIPREngine",
+    "BeliefValidationEngine",
 ]
