+++ utils/__init__.py
@@ -1,4 +1,6 @@
 from importlib import import_module
 import sys
-module = import_module('universalCode.utils')
+
+# TODO: REMOVE after transition to echo_core.utils
+module = import_module('universalCode.echo_core.utils')
 sys.modules[__name__] = module
